strtrim itoa join and first try to warhammer

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: jhidalgo <jhidalgo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/06 18:06:16 by jhidalgo          #+#    #+#             */
-/*   Updated: 2021/04/13 13:34:13 by jhidalgo         ###   ########.fr       */
+/*   Updated: 2021/04/14 17:20:27 by jhidalgo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 char	*ft_strnstr(const char	*big, const char *little, size_t len)
 {
-	int		i;
-	int		j;
-	char	*ptr;
+	size_t		i;
+	size_t		j;
+	char		*ptr;
 
 	ptr = (char *)big;
 	i = 0;
